Make Society.admins a method, add Member.socs
This is a significant API change, which finally prevents get_societies
from reading the memberlist. Society objects are no longer initialised
with a MemberSet of their admins, just with a frozenset of their admin
CRSIDs. The .admins() method will then find a corresponding MemberSet
and cache the result for future lookups. If you happen to have a
dictionary of members handy (say, from members()) then the method can be
given that to do the lookup without reading the memberlist.

The visible change is basically that soc.admins should be replaced with
soc.admins(). Some types of usage may get /much/ faster.

Analogously, Member has a new socs() method.