Require SQLAlchemy <2.0
autocommit=True on SQLAlchemy Sessions was deprecated in 1.4 and removed
in 2.0.  Unfortunately we rely on this heavily, and it is a bit of work
to fix it, so for now constrain to a version that keeps this option.

This doesn't affect current use of the libraries on the SRCF, because we
use Ubuntu's packaged version of SQLAlchemy (1.3 in Focal, and 1.4 in
Jammy once we upgrade to it), but it does affect local dev work and venv
setup using `make venv`.  (It would probably make sense to pin all the
dependencies to versions matching our targeted deployment(s) on Ubuntu.)

While we're touching setup.py, format it a bit more nicely (4-space
indentation rather than hanging indent based on the opening bracket).